HTTP Endpoint documentation (#930)
## Goal
Introduce HTTP API documentation to allow users to use the new TypeDB
HTTP Endpoint starting from TypeDB 3.2.0-rc0.

## Implementation
This is the first version of the docs, and they can be extended and
enhanced in terms of the UX in the near future.

Add:
* HTTP "driver" docs with general description, setup and configuration
guides, version compatibility, and the full API reference.
* Code snippets for basic operations in Manuals: database and user
management, working with transactions.
* New server configuration options in Manual.

Fix:
* Small typos.
* Invalid driver usage examples.